Code clean up after in-person code review. Consolidate update_context, remove superfluous tests.
